requirements.txt
flask==2.2.5
spacy==3.7.1
pdfplumber==0.8.0
python-docx==0.8.11
filetype==1.2.0
fuzzywuzzy==0.18.0
python-Levenshtein==0.21.1
